rename package in demo app for spring
