+++ oidc-spring-support-demo/src/main/java/no/nav/security/token/support/demo/spring/DemoApplication.java
@@ -1,4 +1,4 @@
-package no.nav.security.oidcspringsupportdemo;
+package no.nav.security.token.support.demo.spring;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
