+++ oidc-spring-support-demo/src/main/java/no/nav/security/token/support/demo/spring/config/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package no.nav.security.oidcspringsupportdemo.config;
+package no.nav.security.token.support.demo.spring.config;
 
 import no.nav.security.token.support.spring.api.EnableJwtTokenValidation;
 import org.springframework.context.annotation.Configuration;
